Fix CodeQL: cast before multiply to avoid 32-bit overflow on widening
CodeQL flagged cpp/integer-multiplication-cast-to-long across several
files this PR touched: an unsigned int/int multiplication result was
implicitly converted to a wider type (size_t/int64_t) only after the
multiply, so the multiply itself could overflow before widening for
large tensor dimensions.

- rnn.cpp: timestep * feature_size / timestep * unit offsets into
  getSharedDataTensor (size_t offset param).
- split_layer.h: batch * leading_helper_dim in setBatch.
- zoneout_lstmcell.cpp: max_timestep * batch_size building TensorDim
  (size_t ctor param).
- float_tensor.cpp: i * e.strides[cur_axis], where e.strides is
  array<unsigned int, ...> (the sibling `strides` member is already
  array<size_t, ...>, so only this line needed a cast).
- q4_0_utils.cpp: (K / QK4_0) * (N / 4|8) building data_size, and
  K * N passed to dequantize_row_q4_0's int64_t k parameter.

Fix casts the first operand to the target width so the multiply
itself happens in that width. Verified each translation unit still
compiles clean under -Werror with the project's actual build flags.

+++ nntrainer/layers/rnn.cpp
@@ -200,10 +200,10 @@ void RNNLayer::forwarding(RunLayerContext &context, bool training) {
     Tensor hidden_state_slice = hidden_state.getBatchSlice(batch, 1);
 
     for (unsigned int timestep = 0; timestep < max_timestep; ++timestep) {
-      Tensor in = input_slice.getSharedDataTensor({feature_size},
-                                                  timestep * feature_size);
+      Tensor in = input_slice.getSharedDataTensor(
+        {feature_size}, (size_t)timestep * feature_size);
       Tensor hs =
-        hidden_state_slice.getSharedDataTensor({unit}, timestep * unit);
+        hidden_state_slice.getSharedDataTensor({unit}, (size_t)timestep * unit);
 
       in.dot(weight_ih, hs);
       if (!disable_bias) {
@@ -216,8 +216,8 @@ void RNNLayer::forwarding(RunLayerContext &context, bool training) {
       }
 
       if (timestep) {
-        Tensor prev_hs =
-          hidden_state_slice.getSharedDataTensor({unit}, (timestep - 1) * unit);
+        Tensor prev_hs = hidden_state_slice.getSharedDataTensor(
+          {unit}, (size_t)(timestep - 1) * unit);
         prev_hs.dot(weight_hh, hs, false, false, 1.0);
       }
 
@@ -228,7 +228,7 @@ void RNNLayer::forwarding(RunLayerContext &context, bool training) {
         Tensor dropout_mask = context.getTensor(wt_idx[RNNParams::dropout_mask])
                                 .getBatchSlice(batch, 1);
         Tensor dropout_mask_t =
-          dropout_mask.getSharedDataTensor({unit}, timestep * unit);
+          dropout_mask.getSharedDataTensor({unit}, (size_t)timestep * unit);
         dropout_mask_t.dropout_mask(dropout_rate);
         hs.multiply_i(dropout_mask_t);
       }
